feat(US-05): Adição de todas funcionalidades para BLOQUEIO e DESBLOQUEIO de tarefas
além da refatoração de alguns códigos, foi adicionado a função de desbloquear a tarefa.

+++ task-service/src/main/java/com/kernelpanic/task_service/dtos/BloquearTarefaDTO.java
@@ -1,7 +1,7 @@
 package com.kernelpanic.task_service.dtos;
 
 public class BloquearTarefaDTO {
-    public Long usuarioId;
+    public Integer usuarioId;
     public String categoria;
     public String descricao;
 }
